weather/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-autoslug==1.9.8
django-cities-light==3.9.0
djangorestframework==3.13.1
mysqlclient==2.1.0
progressbar2==4.0.0
python-utils==3.1.0
pytz==2021.3
sqlparse==0.4.2
Unidecode==1.3.3
